test(arceos-ipi): skip aarch64 and split implementation
